# public/assets/api.yaml
openapi: 3.0.0
info:
  title: MAXIMUM
  description: 'This is docs rest api simple blogs'
  contact:
    email: lintanglazuardi11@gmail.com
  version: 1.0.0
servers:
  -
    url: /api
    description: 'Development Server'
paths:
  /login:
    post:
      tags:
        - Auth
      summary: 'User login'
      description: 'Logs in a user using email and password'
      operationId: userLogin
      requestBody:
        description: 'User credentials'
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  example: admin@gmail.com
                password:
                  type: string
                  example: admin
              type: object
      responses:
        '200':
          description: 'Successful login'
          content:
            application/json:
              schema:
                properties:
                  status: { type: integer, example: 200 }
                  error: { type: bool, example: false }
                  message: { type: string, example: 'Login successful' }
                  data: { properties: { token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... } }, type: object }
                type: object
        '404':
          description: 'Invalid username or password'
          content:
            application/json:
              schema:
                properties:
                  status: { type: integer, example: 401 }
                  error: { type: bool, example: true }
                  message: { type: string, example: 'Invalid username or password' }
                  data: { type: array, items: {  } }
                type: object
  /register:
    post:
      tags:
        - Auth
      summary: 'User Registrasi'
      description: 'Registrasi in a user'
      operationId: userRegistrasi
      requestBody:
        description: 'User credentials'
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: admin
                email:
                  type: string
                  example: admin@gmail.com
                password:
                  type: string
                  example: admin
              type: object
      responses:
        '200':
          description: 'User added successfully'
          content:
            application/json:
              schema:
                properties:
                  status: { type: integer, example: 200 }
                  error: { type: bool, example: false }
                  message: { type: string, example: 'User added successfully' }
                  data: { properties: { token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... } }, type: object }
                type: object
        '404':
          description: 'Invalid username or password'
          content:
            application/json:
              schema:
                properties:
                  status: { type: integer, example: 401 }
                  error: { type: bool, example: true }
                  message: { type: string, example: 'Invalid username or password' }
                  data: { type: array, items: {  } }
                type: object
...
components:
...
  securitySchemes:
    bearer_auth:
      type: http
      name: bearer_auth
      bearerFormat: JWT
      scheme: bearer
...